requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
fastapi
uvicorn
streamlit
requests
langchain-groq
langchain-core
langgraph
psycopg2-binary
sqlalchemy
python-dotenv
fastmcp